+++ packages/theme/src/index.ts
@@ -35,8 +35,8 @@ import UserWorksPage from './components/UserWorks.vue'
 // 内置一些特殊的主题色
 import './styles/theme/inline-theme.var.css'
 
-// 导入group icons
-import 'virtual:group-icons.css'
+// 导入group icons，改由自定义插件导入
+// import 'virtual:group-icons.css'
 
 export const BlogTheme: Theme = {
   ...DefaultTheme,